+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `classify-train` command now takes `--label-format` argument.
 - `classify-predict` command now takes `--output-format` argument.
+- `classify-predict` command now takes `--label-type` argument.
 
 ## [1.0.1] - 2026-03-19